Double cast

+++ src/drivers/lis3mdl/lis3mdl.cpp
@@ -1152,12 +1152,12 @@ int LIS3MDL::calibrate(struct file *filp, unsigned enable)
 	mscale_previous.z_scale = 1.0f / scaling[2];
 	
 	PX4_INFO("******* Value of mscale_previous.x_scale : %f",(double)mscale_previous.x_scale);
-	PX4_INFO("******* Value of mscale_previous.y_scale : %f",(float)mscale_previous.y_scale);
-	PX4_INFO("******* Value of mscale_previous.z_scale : %f",(float)mscale_previous.z_scale);
+	PX4_INFO("******* Value of mscale_previous.y_scale : %f",(double)mscale_previous.y_scale);
+	PX4_INFO("******* Value of mscale_previous.z_scale : %f",(double)mscale_previous.z_scale);
 
-	PX4_INFO("******* Value of _scale.x_scale : %f",(float)_scale.x_scale);
-	PX4_INFO("******* Value of _scale.y_scale : %f",(float)_scale.y_scale);
-	PX4_INFO("******* Value of _scale.z_scale : %f",(float)_scale.z_scale);
+	PX4_INFO("******* Value of _scale.x_scale : %f",(double)_scale.x_scale);
+	PX4_INFO("******* Value of _scale.y_scale : %f",(double)_scale.y_scale);
+	PX4_INFO("******* Value of _scale.z_scale : %f",(double)_scale.z_scale);
 	
 	ret = OK;
 
@@ -1167,9 +1167,9 @@ out:
 		warn("FAILED: MAGIOCSSCALE 2");
 	}
 	
-	PX4_INFO("******* Value of _scale.x_scale : %f",(float)_scale.x_scale);
-	PX4_INFO("******* Value of _scale.y_scale : %f",(float)_scale.y_scale);
-	PX4_INFO("******* Value of _scale.z_scale : %f",(float)_scale.z_scale);
+	PX4_INFO("******* Value of _scale.x_scale : %f",(double)_scale.x_scale);
+	PX4_INFO("******* Value of _scale.y_scale : %f",(double)_scale.y_scale);
+	PX4_INFO("******* Value of _scale.z_scale : %f",(double)_scale.z_scale);
 
 	/* set back to normal mode */
 	/* Set to 4 Gauss */
